requirements.txt
requests
uptime-kuma-api
pandas
python-dotenv